Split activity integration test ownership
