+++ Tests/QuillCodeParityTests/ParityWorkspaceModelActivityGateTests.swift
@@ -7,26 +7,103 @@ final class ParityWorkspaceModelActivityGateTests: QuillCodeParityTestCase {
         let retryPlannerText = try Self.appSourceText(named: "WorkspaceRetryPlanner.swift")
         let retryPlannerTests = try Self.appTestSourceText(named: "WorkspaceRetryPlannerTests.swift")
 
-        XCTAssertTrue(retryPlannerText.contains("enum WorkspaceRetryPlanner"), "Retry planning should live in a focused helper.")
-        XCTAssertTrue(retryPlannerText.contains("static func canRetryLastUserTurn"), "Retry availability should be directly testable.")
-        XCTAssertTrue(retryPlannerText.contains("static func retryDraft"), "Retry draft selection should be directly testable.")
-        XCTAssertTrue(composerText.contains("WorkspaceRetryPlanner.canRetryLastUserTurn"), "WorkspaceModel composer APIs should delegate retry availability.")
-        XCTAssertTrue(composerText.contains("WorkspaceRetryPlanner.retryDraft"), "WorkspaceModel composer APIs should delegate retry draft selection.")
-        XCTAssertTrue(retryPlannerTests.contains("testRetryDraftUsesLatestNonEmptyUserMessageAndPreservesOriginalText"), "Retry draft behavior should have focused coverage.")
-        XCTAssertTrue(retryPlannerTests.contains("testRetryRequiresUserMessageAndIdleComposer"), "Retry availability should have focused coverage.")
-        XCTAssertFalse(modelText.contains("messages.last(where:"), "WorkspaceModel should not scan transcript messages for retry drafts.")
-        XCTAssertFalse(modelText.contains("messages.contains {"), "WorkspaceModel should not own retry availability scans.")
-        XCTAssertFalse(modelText.contains("WorkspaceRetryPlanner.canRetryLastUserTurn"), "WorkspaceModel.swift should not own retry availability APIs.")
+        assertContains(
+            retryPlannerText,
+            "enum WorkspaceRetryPlanner",
+            "Retry planning should live in a focused helper."
+        )
+        assertContains(retryPlannerText, "static func canRetryLastUserTurn", "Retry availability should be testable.")
+        assertContains(retryPlannerText, "static func retryDraft", "Retry draft selection should be testable.")
+        assertContains(
+            composerText,
+            "WorkspaceRetryPlanner.canRetryLastUserTurn",
+            "Composer should delegate retry availability."
+        )
+        assertContains(
+            composerText,
+            "WorkspaceRetryPlanner.retryDraft",
+            "Composer should delegate retry draft selection."
+        )
+        assertContains(
+            retryPlannerTests,
+            "testRetryDraftUsesLatestNonEmptyUserMessageAndPreservesOriginalText",
+            "Retry draft behavior should have focused coverage."
+        )
+        assertContains(
+            retryPlannerTests,
+            "testRetryRequiresUserMessageAndIdleComposer",
+            "Retry availability should have focused coverage."
+        )
+        assertExcludes(modelText, "messages.last(where:", "WorkspaceModel should not scan transcript messages.")
+        assertExcludes(modelText, "messages.contains {", "WorkspaceModel should not own retry availability scans.")
+        assertExcludes(
+            modelText,
+            "WorkspaceRetryPlanner.canRetryLastUserTurn",
+            "WorkspaceModel.swift should not own retry availability APIs."
+        )
     }
 
     func testWorkspaceActivityIntegrationTestsOwnModelActivityFlows() throws {
         let modelTests = try Self.appTestSourceText(named: "WorkspaceModelTests.swift")
         let activityIntegrationTests = try Self.appTestSourceText(named: "WorkspaceActivityIntegrationTests.swift")
+        let instructionTests = try Self.appTestSourceText(named: "WorkspaceActivityInstructionIntegrationTests.swift")
+        let planHandoffTests = try Self.appTestSourceText(named: "WorkspaceActivityPlanHandoffIntegrationTests.swift")
 
-        XCTAssertTrue(activityIntegrationTests.contains("testPlanUpdateToolRecordsNormalizedActivityPlan"), "Plan-update activity integration should live in focused activity tests.")
-        XCTAssertTrue(activityIntegrationTests.contains("testPlanUpdateToolRejectsMultipleRunningSteps"), "Plan-update rejection integration should live in focused activity tests.")
-        XCTAssertFalse(modelTests.contains("testPlanUpdateToolRecordsNormalizedActivityPlan"), "WorkspaceModelTests should not own plan-update activity integration flows.")
-        XCTAssertFalse(modelTests.contains("testPlanUpdateToolRejectsMultipleRunningSteps"), "WorkspaceModelTests should not own plan-update rejection integration flows.")
+        assertContains(
+            instructionTests,
+            "testActivitySourcesSurfaceInstructionDiagnostics",
+            "Instruction source coverage should stay in focused instruction activity tests."
+        )
+        assertContains(
+            instructionTests,
+            "testActivitySourcesSurfaceInstructionSemanticConflictDiagnostics",
+            "Instruction-review coverage should stay in focused instruction activity tests."
+        )
+        assertContains(
+            planHandoffTests,
+            "testPlanUpdateToolRecordsNormalizedActivityPlan",
+            "Plan-update coverage should stay in focused plan/handoff activity tests."
+        )
+        assertContains(
+            planHandoffTests,
+            "testPlanUpdateToolRejectsMultipleRunningSteps",
+            "Plan-update rejection coverage should stay in focused plan/handoff activity tests."
+        )
+        assertExcludes(
+            activityIntegrationTests,
+            "testActivitySourcesSurfaceInstructionDiagnostics",
+            "Surface/context activity tests should not own instruction source flows."
+        )
+        assertExcludes(
+            activityIntegrationTests,
+            "testActivitySourcesSurfaceInstructionSemanticConflictDiagnostics",
+            "Surface/context activity tests should not own instruction-review flows."
+        )
+        assertExcludes(
+            activityIntegrationTests,
+            "testPlanUpdateToolRecordsNormalizedActivityPlan",
+            "Surface/context activity tests should not own plan-update tool flows."
+        )
+        assertExcludes(
+            activityIntegrationTests,
+            "testPlanUpdateToolRejectsMultipleRunningSteps",
+            "Surface/context activity tests should not own plan-update rejection flows."
+        )
+        assertExcludes(
+            modelTests,
+            "testActivitySourcesSurfaceInstructionDiagnostics",
+            "WorkspaceModelTests should not own instruction source activity flows."
+        )
+        assertExcludes(
+            modelTests,
+            "testPlanUpdateToolRecordsNormalizedActivityPlan",
+            "WorkspaceModelTests should not own plan-update activity flows."
+        )
+        assertExcludes(
+            modelTests,
+            "testPlanUpdateToolRejectsMultipleRunningSteps",
+            "WorkspaceModelTests should not own plan-update rejection flows."
+        )
     }
 
     func testWorkspaceActivitySurfaceUsesFocusedBuilderAndSectionTypes() throws {
@@ -40,61 +117,145 @@ final class ParityWorkspaceModelActivityGateTests: QuillCodeParityTestCase {
         let textHelperText = try Self.appSourceText(named: "WorkspaceActivityText.swift")
         let statusText = try Self.appSourceText(named: "WorkspaceActivityStatusLabel.swift")
 
-        XCTAssertTrue(surfaceText.contains("public struct WorkspaceActivitySurface"), "Activity surface payload should keep the public DTO entry point.")
-        XCTAssertTrue(surfaceText.contains("WorkspaceActivitySurfaceBuilder.sections"), "Activity surface should delegate section construction.")
-        XCTAssertTrue(surfaceText.contains("WorkspaceActivitySurfaceBuilder.planItems"), "Activity surface should delegate derived task-plan rows.")
-        XCTAssertTrue(builderText.contains("enum WorkspaceActivitySurfaceBuilder"), "Activity derivation should live in a focused builder.")
-        XCTAssertTrue(builderText.contains("WorkspaceActivityPlanSurfaceBuilder.fallbackItems"), "Activity builder should delegate fallback plan rows.")
-        XCTAssertTrue(builderText.contains("WorkspaceActivityPlanSurfaceBuilder.authoredItems"), "Activity builder should delegate authored plan rows.")
-        XCTAssertTrue(builderText.contains("WorkspaceActivityEventSurfaceBuilder.recentSteps"), "Activity builder should delegate event-row projection.")
-        XCTAssertTrue(builderText.contains("WorkspaceActivitySourceSurfaceBuilder.items"), "Activity builder should delegate source-row projection.")
-        XCTAssertTrue(builderText.contains("instructionConflictItems"), "Activity builder should promote instruction conflicts into a first-class review section.")
-        XCTAssertTrue(builderText.contains("WorkspaceActivityHandoffSummaryBuilder.summary"), "Activity builder should delegate handoff summary copy.")
-        XCTAssertTrue(planBuilderText.contains("enum WorkspaceActivityPlanSurfaceBuilder"), "Plan-row derivation should have a focused owner.")
-        XCTAssertTrue(planBuilderText.contains("static func authoredItems"), "Authored plan rows should stay beside fallback plan rows.")
-        XCTAssertTrue(planBuilderText.contains("private static func reviewState"), "Fallback plan review-state copy should stay in the plan builder.")
-        XCTAssertTrue(eventBuilderText.contains("enum WorkspaceActivityEventSurfaceBuilder"), "Event-row projection should have a focused owner.")
-        XCTAssertTrue(eventBuilderText.contains("private static func eventKindLabel"), "Event labeling should stay beside event-row projection.")
-        XCTAssertTrue(sourceBuilderText.contains("enum WorkspaceActivitySourceSurfaceBuilder"), "Instruction and memory source rows should have a focused owner.")
-        XCTAssertTrue(handoffBuilderText.contains("enum WorkspaceActivityHandoffSummaryBuilder"), "Handoff summary copy should have a focused owner.")
-        XCTAssertTrue(textHelperText.contains("enum WorkspaceActivityText"), "Shared activity text formatting should not be copied between builders.")
-        XCTAssertTrue(statusText.contains("enum ActivityStatusLabel"), "Activity status labels should be shared by focused builders.")
-        XCTAssertTrue(sectionText.contains("public enum ActivitySectionKind"), "Activity section metadata should live beside section DTOs.")
-        XCTAssertTrue(sectionText.contains("case instructionReview"), "Instruction conflict review should be a first-class Activity section.")
-        XCTAssertTrue(sectionText.contains("public struct ActivitySectionSurface"), "Activity section DTOs should live outside the root surface file.")
-        XCTAssertTrue(sectionText.contains("public struct ActivityItemSurface"), "Activity item DTOs should live outside the root surface file.")
-        XCTAssertFalse(surfaceText.contains("private static func planItems"), "Activity surface should not own plan derivation.")
-        XCTAssertFalse(surfaceText.contains("private static func recentSteps"), "Activity surface should not own event-row derivation.")
-        XCTAssertFalse(surfaceText.contains("public enum ActivitySectionKind"), "Activity surface should not own section metadata.")
-        XCTAssertFalse(builderText.contains("private static func eventKindLabel"), "Top-level activity builder should not own event labeling.")
-        XCTAssertFalse(builderText.contains("private static func reviewState"), "Top-level activity builder should not own fallback plan review state.")
-        XCTAssertFalse(builderText.contains("Latest answer:"), "Top-level activity builder should not own handoff summary prose.")
+        assertContains(
+            surfaceText,
+            "public struct WorkspaceActivitySurface",
+            "Activity surface should keep the public DTO entry point."
+        )
+        assertContains(
+            surfaceText,
+            "WorkspaceActivitySurfaceBuilder.sections",
+            "Activity surface should delegate section construction."
+        )
+        assertContains(
+            surfaceText,
+            "WorkspaceActivitySurfaceBuilder.planItems",
+            "Activity surface should delegate task-plan row derivation."
+        )
+        assertContains(builderText, "enum WorkspaceActivitySurfaceBuilder", "Activity derivation should be focused.")
+        assertContains(
+            builderText,
+            "WorkspaceActivityPlanSurfaceBuilder.fallbackItems",
+            "Fallback rows should delegate."
+        )
+        assertContains(
+            builderText,
+            "WorkspaceActivityPlanSurfaceBuilder.authoredItems",
+            "Authored rows should delegate."
+        )
+        assertContains(builderText, "WorkspaceActivityEventSurfaceBuilder.recentSteps", "Event rows should delegate.")
+        assertContains(builderText, "WorkspaceActivitySourceSurfaceBuilder.items", "Source rows should delegate.")
+        assertContains(builderText, "instructionConflictItems", "Instruction conflicts should be first-class.")
+        assertContains(builderText, "WorkspaceActivityHandoffSummaryBuilder.summary", "Handoff copy should delegate.")
+        assertContains(planBuilderText, "enum WorkspaceActivityPlanSurfaceBuilder", "Plan rows need a focused owner.")
+        assertContains(
+            planBuilderText,
+            "static func authoredItems",
+            "Authored plan rows should stay with fallback rows."
+        )
+        assertContains(planBuilderText, "private static func reviewState", "Fallback review-state copy belongs here.")
+        assertContains(
+            eventBuilderText,
+            "enum WorkspaceActivityEventSurfaceBuilder",
+            "Event rows need a focused owner."
+        )
+        assertContains(
+            eventBuilderText,
+            "private static func eventKindLabel",
+            "Event labeling belongs with event rows."
+        )
+        assertContains(
+            sourceBuilderText,
+            "enum WorkspaceActivitySourceSurfaceBuilder",
+            "Source rows need a focused owner."
+        )
+        assertContains(
+            handoffBuilderText,
+            "enum WorkspaceActivityHandoffSummaryBuilder",
+            "Handoff copy needs a focused owner."
+        )
+        assertContains(textHelperText, "enum WorkspaceActivityText", "Shared activity text should not be copied.")
+        assertContains(statusText, "enum ActivityStatusLabel", "Activity status labels should be shared.")
+        assertContains(
+            sectionText,
+            "public enum ActivitySectionKind",
+            "Section metadata should live beside section DTOs."
+        )
+        assertContains(sectionText, "case instructionReview", "Instruction review should be a first-class section.")
+        assertContains(
+            sectionText,
+            "public struct ActivitySectionSurface",
+            "Activity section DTOs should live outside the root surface file."
+        )
+        assertContains(
+            sectionText,
+            "public struct ActivityItemSurface",
+            "Activity item DTOs should live outside the root surface file."
+        )
+        assertExcludes(surfaceText, "private static func planItems", "Activity surface should not own plan derivation.")
+        assertExcludes(surfaceText, "private static func recentSteps", "Activity surface should not own event rows.")
+        assertExcludes(surfaceText, "public enum ActivitySectionKind", "Activity surface should not own sections.")
+        assertExcludes(builderText, "private static func eventKindLabel", "Top-level builder should not own labels.")
+        assertExcludes(builderText, "private static func reviewState", "Top-level builder should not own review state.")
+        assertExcludes(builderText, "Latest answer:", "Top-level builder should not own handoff prose.")
     }
 
     func testWorkspaceToolCardIntegrationTestsOwnModelToolCardFlows() throws {
         let modelTests = try Self.appTestSourceText(named: "WorkspaceModelTests.swift")
         let toolCardIntegrationTests = try Self.appTestSourceText(named: "WorkspaceToolCardIntegrationTests.swift")
 
-        XCTAssertTrue(toolCardIntegrationTests.contains("testToolCardsRepresentActionableApprovalReview"), "Tool-card review projection should live in focused tool-card integration tests.")
-        XCTAssertTrue(toolCardIntegrationTests.contains("testToolCardApprovalActionRecordsDecisionAndRunsTool"), "Tool-card approval execution should live in focused tool-card integration tests.")
-        XCTAssertTrue(toolCardIntegrationTests.contains("testToolCardsRepresentStoppedActiveToolAsFailed"), "Stopped tool-card projection should live in focused tool-card integration tests.")
-        XCTAssertFalse(modelTests.contains("testToolCardsRepresentActionableApprovalReview"), "WorkspaceModelTests should not own actionable approval-card projection.")
-        XCTAssertFalse(modelTests.contains("testToolCardApprovalActionRecordsDecisionAndRunsTool"), "WorkspaceModelTests should not own approval-card execution integration.")
-        XCTAssertFalse(modelTests.contains("testToolCardsRepresentStoppedActiveToolAsFailed"), "WorkspaceModelTests should not own stopped tool-card projection.")
+        assertContains(
+            toolCardIntegrationTests,
+            "testToolCardsRepresentActionableApprovalReview",
+            "Tool-card review projection should live in focused integration tests."
+        )
+        assertContains(
+            toolCardIntegrationTests,
+            "testToolCardApprovalActionRecordsDecisionAndRunsTool",
+            "Tool-card approval execution should live in focused integration tests."
+        )
+        assertContains(
+            toolCardIntegrationTests,
+            "testToolCardsRepresentStoppedActiveToolAsFailed",
+            "Stopped tool-card projection should live in focused integration tests."
+        )
+        assertExcludes(
+            modelTests,
+            "testToolCardsRepresentActionableApprovalReview",
+            "WorkspaceModelTests should not own approval-card projection."
+        )
+        assertExcludes(
+            modelTests,
+            "testToolCardApprovalActionRecordsDecisionAndRunsTool",
+            "WorkspaceModelTests should not own approval-card execution."
+        )
+        assertExcludes(
+            modelTests,
+            "testToolCardsRepresentStoppedActiveToolAsFailed",
+            "WorkspaceModelTests should not own stopped tool-card projection."
+        )
     }
 
     func testWorkspaceModelTestsRemainRetired() throws {
         let modelTests = try Self.appTestSourceText(named: "WorkspaceModelTests.swift")
 
-        XCTAssertTrue(modelTests.contains("Intentionally empty"), "WorkspaceModelTests should stay as a visible retirement marker.")
-        XCTAssertFalse(modelTests.contains("func test"), "New workspace integration coverage should use a focused feature test suite, not WorkspaceModelTests.")
+        assertContains(modelTests, "Intentionally empty", "WorkspaceModelTests should stay as a retirement marker.")
+        assertExcludes(modelTests, "func test", "New workspace coverage should use focused feature test suites.")
     }
 
     func testFocusedWorkspaceUnitSuitesUseSharedTemporaryDirectorySupport() throws {
         let supportText = try Self.appTestSourceText(named: "WorkspaceModelIntegrationTestSupport.swift")
-        XCTAssertTrue(supportText.contains("extension XCTestCase"), "App integration temp helpers should live on XCTestCase so they can register teardown cleanup.")
-        XCTAssertTrue(supportText.contains("func makeTempDirectory() throws -> URL"), "Legacy app integration tests should route through the shared temp-directory wrapper.")
-        XCTAssertTrue(supportText.contains("makeQuillCodeTestDirectory()"), "App integration temp helpers should delegate to the teardown-backed helper.")
+        assertContains(supportText, "extension XCTestCase", "Temp helpers should register teardown cleanup.")
+        assertContains(
+            supportText,
+            "func makeTempDirectory() throws -> URL",
+            "Legacy app integration tests should use the shared wrapper."
+        )
+        assertContains(
+            supportText,
+            "makeQuillCodeTestDirectory()",
+            "App integration temp helpers should delegate to teardown-backed support."
+        )
 
         let suiteNames = [
             "WorkspaceAgentRunContextBuilderTests.swift",
@@ -107,16 +268,19 @@ final class ParityWorkspaceModelActivityGateTests: QuillCodeParityTestCase {
 
         for suiteName in suiteNames {
             let suiteText = try Self.appTestSourceText(named: suiteName)
-            XCTAssertTrue(
-                suiteText.contains("makeQuillCodeTestDirectory()"),
+            assertContains(
+                suiteText,
+                "makeQuillCodeTestDirectory()",
                 "\(suiteName) should use the shared teardown-backed test directory helper."
             )
-            XCTAssertFalse(
-                suiteText.contains("private func temporaryDirectory"),
+            assertExcludes(
+                suiteText,
+                "private func temporaryDirectory",
                 "\(suiteName) should not reintroduce a private temp-directory helper."
             )
-            XCTAssertFalse(
-                suiteText.contains("FileManager.default.temporaryDirectory.appendingPathComponent(UUID().uuidString)"),
+            assertExcludes(
+                suiteText,
+                "FileManager.default.temporaryDirectory.appendingPathComponent(UUID().uuidString)",
                 "\(suiteName) should not build untracked temp directories inline."
             )
         }
@@ -131,19 +295,42 @@ final class ParityWorkspaceModelActivityGateTests: QuillCodeParityTestCase {
         ]
         for suiteName in integrationSuiteNames {
             let suiteText = try Self.appTestSourceText(named: suiteName)
-            XCTAssertFalse(
-                suiteText.contains("private func makeTempDirectory()"),
+            assertExcludes(
+                suiteText,
+                "private func makeTempDirectory()",
                 "\(suiteName) should use WorkspaceModelIntegrationTestSupport.makeTempDirectory()."
             )
-            XCTAssertFalse(
-                suiteText.contains("NSTemporaryDirectory()"),
+            assertExcludes(
+                suiteText,
+                "NSTemporaryDirectory()",
                 "\(suiteName) should not build untracked temp directories inline."
             )
-            XCTAssertFalse(
-                suiteText.contains("FileManager.default.temporaryDirectory"),
+            assertExcludes(
+                suiteText,
+                "FileManager.default.temporaryDirectory",
                 "\(suiteName) should not build untracked temp directories inline."
             )
         }
     }
 
+}
+
+private func assertContains(
+    _ text: String,
+    _ needle: String,
+    _ message: String,
+    file: StaticString = #filePath,
+    line: UInt = #line
+) {
+    XCTAssertTrue(text.contains(needle), message, file: file, line: line)
+}
+
+private func assertExcludes(
+    _ text: String,
+    _ needle: String,
+    _ message: String,
+    file: StaticString = #filePath,
+    line: UInt = #line
+) {
+    XCTAssertFalse(text.contains(needle), message, file: file, line: line)
 }
